backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
django-filter==24.3
Pillow==11.1.0
python-dotenv==1.0.1
google-generativeai==0.8.4
